// src/app/table-list/model/encuesta.model.ts
export interface Encuesta {
    id: number;
    pregunta1: string[];
    pregunta2: string[] | null;
    pregunta3: string[] | null;
    pregunta4: string;
    pregunta5: string[] | null;
    pregunta6: string;
    pregunta7: string[];
    pregunta8: string | null;
    pregunta9: string;
    created_at: string;
    

}